+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.26.0] - 2018-08-20
+### Added
+- Basic documentation for gen-sample commands
+
+### Changed
+- Update Stacker sample module with Terraform-supporting template
+
 ## [0.25.0] - 2018-08-17
 ### Added
 - Allow environments to be specified at top level of deployment
@@ -249,7 +256,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Fix changed CFN parameters not being displayed during `runway plan`.
 
-[Unreleased]: https://github.com/onicagroup/runway/compare/v0.25.0...HEAD
+[Unreleased]: https://github.com/onicagroup/runway/compare/v0.26.0...HEAD
+[0.26.0]: https://github.com/onicagroup/runway/compare/v0.25.0...v0.26.0
 [0.25.0]: https://github.com/onicagroup/runway/compare/v0.24.0...v0.25.0
 [0.24.0]: https://github.com/onicagroup/runway/compare/v0.23.3...v0.24.0
 [0.23.3]: https://github.com/onicagroup/runway/compare/v0.23.2...v0.23.3